feat(docs-and-either): move either to application and add docs images
