+++ application/src/main/java/com/kaua/monitoring/application/either/Either.java
@@ -1,4 +1,4 @@
-package com.kaua.monitoring.application.exceptions.either;
+package com.kaua.monitoring.application.either;
 
 import java.util.NoSuchElementException;
