+++ application/src/main/java/com/kaua/monitoring/application/usecases/link/create/CreateLinkUseCase.java
@@ -1,7 +1,7 @@
 package com.kaua.monitoring.application.usecases.link.create;
 
 import com.kaua.monitoring.application.exceptions.DomainException;
-import com.kaua.monitoring.application.exceptions.either.Either;
+import com.kaua.monitoring.application.either.Either;
 import com.kaua.monitoring.application.usecases.UseCase;
 import com.kaua.monitoring.application.usecases.link.outputs.CreateLinkOutput;
