+++ application/src/main/java/com/kaua/monitoring/application/usecases/link/create/DefaultCreateLinkUseCase.java
@@ -2,7 +2,7 @@ package com.kaua.monitoring.application.usecases.link.create;
 
 import com.kaua.monitoring.application.exceptions.DomainException;
 import com.kaua.monitoring.application.exceptions.NotFoundException;
-import com.kaua.monitoring.application.exceptions.either.Either;
+import com.kaua.monitoring.application.either.Either;
 import com.kaua.monitoring.application.gateways.LinkGateway;
 import com.kaua.monitoring.application.gateways.ProfileGateway;
 import com.kaua.monitoring.application.usecases.link.outputs.CreateLinkOutput;
