+++ application/src/main/java/com/kaua/monitoring/application/usecases/profile/create/CreateProfileUseCase.java
@@ -1,7 +1,7 @@
 package com.kaua.monitoring.application.usecases.profile.create;
 
 import com.kaua.monitoring.application.exceptions.DomainException;
-import com.kaua.monitoring.application.exceptions.either.Either;
+import com.kaua.monitoring.application.either.Either;
 import com.kaua.monitoring.application.usecases.UseCase;
 import com.kaua.monitoring.application.usecases.profile.outputs.CreateProfileOutput;
